Clean-up POMs and IT. Run IT with Java 6.
