+++ pom.xml
@@ -84,6 +84,7 @@
         <jdk>[1.6,)</jdk>
       </activation>
       <modules>
+        <module>xstream-its</module>
         <module>xstream-jmh</module>
         <module>xstream-distribution</module>
       </modules>
@@ -263,7 +264,6 @@
     <module>xstream</module>
     <module>xstream-hibernate</module>
     <module>xstream-benchmark</module>
-    <module>xstream-its</module>
   </modules>
 
   <licenses>
@@ -620,66 +620,57 @@
 
       <!-- Pax Exam Dependencies -->
       <dependency>
-        <groupId>org.ops4j.pax.exam</groupId>
-        <artifactId>pax-exam-junit4</artifactId>
-        <version>${version.pax.exam}</version>
+        <groupId>org.apache.felix</groupId>
+        <artifactId>org.apache.felix.framework</artifactId>
+        <version>${version.org.apache.felix}</version>
         <scope>test</scope>
       </dependency>
-
       <dependency>
-        <groupId>org.ops4j.pax.exam</groupId>
-        <artifactId>pax-exam-invoker-junit</artifactId>
-        <version>${version.pax.exam}</version>
-        <scope>test</scope>
+        <groupId>org.glassfish.hk2.external</groupId>
+        <artifactId>javax.inject</artifactId>
+        <version>${version.javax.inject}</version>
+        <scope>provided</scope>
       </dependency>
-
       <dependency>
         <groupId>org.ops4j.pax.exam</groupId>
         <artifactId>pax-exam-container-native</artifactId>
-        <version>${version.pax.exam}</version>
+        <version>${version.org.ops4j.pax.exam}</version>
         <scope>test</scope>
       </dependency>
-
-      <dependency>
-        <groupId>org.ops4j.pax.exam</groupId>
-        <artifactId>pax-exam-inject</artifactId>
-        <version>${version.pax.exam}</version>
-        <scope>test</scope>
-      </dependency>
-
       <dependency>
         <groupId>org.ops4j.pax.exam</groupId>
         <artifactId>pax-exam-extender-service</artifactId>
-        <version>${version.pax.exam}</version>
+        <version>${version.org.ops4j.pax.exam}</version>
         <scope>test</scope>
       </dependency>
-
-      <!-- Preferred link because it does not require an mvn url handler implicitely. -->
       <dependency>
         <groupId>org.ops4j.pax.exam</groupId>
-        <artifactId>pax-exam-link-mvn</artifactId>
-        <version>${version.pax.exam}</version>
+        <artifactId>pax-exam-inject</artifactId>
+        <version>${version.org.ops4j.pax.exam}</version>
         <scope>test</scope>
       </dependency>
-
       <dependency>
-        <groupId>org.apache.geronimo.specs</groupId>
-        <artifactId>geronimo-atinject_1.0_spec</artifactId>
-        <version>1.0</version>
-        <scope>provided</scope>
+        <groupId>org.ops4j.pax.exam</groupId>
+        <artifactId>pax-exam-invoker-junit</artifactId>
+        <version>${version.org.ops4j.pax.exam}</version>
+        <scope>test</scope>
       </dependency>
-
+      <dependency>
+        <groupId>org.ops4j.pax.exam</groupId>
+        <artifactId>pax-exam-junit4</artifactId>
+        <version>${version.org.ops4j.pax.exam}</version>
+        <scope>test</scope>
+      </dependency>
       <dependency>
         <groupId>org.ops4j.pax.exam</groupId>
         <artifactId>pax-exam-link-assembly</artifactId>
-        <version>${version.pax.exam}</version>
+        <version>${version.org.ops4j.pax.exam}</version>
         <scope>test</scope>
       </dependency>
-
       <dependency>
-        <groupId>org.apache.felix</groupId>
-        <artifactId>org.apache.felix.framework</artifactId>
-        <version>${version.felix}</version>
+        <groupId>org.ops4j.pax.exam</groupId>
+        <artifactId>pax-exam-link-mvn</artifactId><!-- Preferred link because it does not require an mvn url handler implicitely. -->
+        <version>${version.org.ops4j.pax.exam}</version>
         <scope>test</scope>
       </dependency>
     </dependencies>
@@ -755,6 +746,11 @@
           <groupId>org.apache.maven.plugins</groupId>
           <artifactId>maven-enforcer-plugin</artifactId>
           <version>${version.plugin.maven.enforcer}</version>
+        </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-failsafe-plugin</artifactId>
+          <version>${version.plugin.maven.failsafe}</version>
         </plugin>
         <plugin>
           <groupId>org.apache.maven.plugins</groupId>
@@ -952,10 +948,6 @@
           <artifactId>jacoco-maven-plugin</artifactId>
           <version>${version.plugin.jacoco}</version>
         </plugin>
-        <plugin>
-          <artifactId>maven-failsafe-plugin</artifactId>
-          <version>2.22.0</version>
-        </plugin>
       </plugins>
     </pluginManagement>
     <plugins>
@@ -1025,6 +1017,7 @@
     <version.plugin.maven.dependency>2.1</version.plugin.maven.dependency>
     <version.plugin.maven.deploy>2.3</version.plugin.maven.deploy>
     <version.plugin.maven.enforcer>1.4</version.plugin.maven.enforcer>
+    <version.plugin.maven.failsafe>2.22.0</version.plugin.maven.failsafe>
     <version.plugin.maven.gpg>1.4</version.plugin.maven.gpg>
     <version.plugin.maven.install>2.2</version.plugin.maven.install>
     <version.plugin.maven.jar>2.2</version.plugin.maven.jar>
@@ -1047,16 +1040,17 @@
     <version.commons.io>1.4</version.commons.io>
     <version.commons.lang>2.4</version.commons.lang>
     <version.dom4j>1.6.1</version.dom4j>
-    <version.felix>4.4.1</version.felix>
     <version.hsqldb>2.2.8</version.hsqldb>
     <version.javaassist>3.12.1.GA</version.javaassist>
     <version.javax.activation>1.1.1</version.javax.activation>
     <version.javax.annotation.api>1.3.2</version.javax.annotation.api>
+    <version.javax.inject>2.4.0</version.javax.inject>
     <version.javax.xml.bind.api>2.3.1</version.javax.xml.bind.api>
     <version.jmock>1.0.1</version.jmock>
     <version.joda-time>1.6</version.joda-time>
     <version.junit>3.8.1</version.junit>
     <version.net.sf.kxml.kxml2>2.3.0</version.net.sf.kxml.kxml2>
+    <version.org.apache.felix>4.4.1</version.org.apache.felix>
     <version.org.codehaus.jettison>1.2</version.org.codehaus.jettison>
     <version.org.codehaus.woodstox.asl>3.2.7</version.org.codehaus.woodstox.asl>
     <version.org.hibernate.core>4.2.5.Final</version.org.hibernate.core>
@@ -1065,8 +1059,8 @@
     <version.org.jdom2>2.0.5</version.org.jdom2>
     <version.org.json>20080701</version.org.json>
     <version.org.openjdk.jmh>1.21</version.org.openjdk.jmh>
+    <version.org.ops4j.pax.exam>3.5.0</version.org.ops4j.pax.exam><!--  java 6 -->
     <version.org.slf4j>1.6.1</version.org.slf4j>
-    <version.pax.exam>4.4.0</version.pax.exam>
     <version.stax>1.2.0</version.stax>
     <version.stax.api>1.0.1</version.stax.api>
     <version.xerces.impl>2.8.1</version.xerces.impl>
