+++ xstream-distribution/src/content/changes.html
@@ -1,7 +1,7 @@
 <html>
 <!--
  Copyright (C) 2005, 2006 Joe Walnes.
- Copyright (C) 2006, 2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018 XStream committers.
+ Copyright (C) 2006, 2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019 XStream committers.
  All rights reserved.
  
  The software in this package is published under the terms of the BSD
@@ -83,6 +83,7 @@
 		<li>Old BEA reference implementation of StAX is outdated, unmaintained and has security issues, therefore
 		XStream's driver has been deprecated.</li>
 		<li>Support for JaCoCo: FieldDictionary ignores synthetic fields starting with <em>$jacoco</em> as name.</li>
+		<li>Add integration test for OSGi (by Wes Wannemacher).</li>
 	</ul>
 
 	<h2>Stream compatibility</h2>
