+++ xstream-distribution/src/content/team.html
@@ -1,7 +1,7 @@
 <html>
 <!--
  Copyright (C) 2005, 2006 Joe Walnes.
- Copyright (C) 2006, 2007, 2008, 2009, 2010, 2011, 2013, 2014, 2015, 2016, 2017 XStream committers.
+ Copyright (C) 2006, 2007, 2008, 2009, 2010, 2011, 2013, 2014, 2015, 2016, 2017, 2019 XStream committers.
  All rights reserved.
  
  The software in this package is published under the terms of the BSD
@@ -125,6 +125,7 @@
       <li>Carsten Hammer</li>
       <li>Matej Cimbora</li>
       <li>Baptiste Mesta</li>
+      <li>Wes Wannemacher</li>
     </ul>
 
     <p>Please direct all correspondence about XStream to the <a href="mailing-lists.html">mailing list</a>
